submit arcsumitb2019_b

+++ arc/sumitb2019/b.cpp
@@ -7,11 +7,11 @@ int main() {
   cin.tie(nullptr);
   cout << fixed << setprecision(15);
 
-  double n;
+  int n;
   cin >> n;
 
   int x = (n + (1.08 - 0.01)) / 1.08;
-  if ((int)n == (int)(x * 1.08)) {
+  if (n == (int)(x * 1.08)) {
     cout << x << '\n';
   } else {
     cout << ":(" << '\n';
